Made minor change to get_hash_stats so the avg_num is a float
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@635 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ vpr/SRC/util/hash.c
@@ -189,7 +189,7 @@ void get_hash_stats(struct s_hash **hash_table, char *hash_table_name){
 	total_elements = total_elements + curr_num;
 	}
 
-	avg_num = total_elements / (HASHSIZE - num_NULL);
+	avg_num = (float) total_elements / ((float)HASHSIZE - (float)num_NULL);
 	
-	vpr_printf(TIO_MESSAGE_INFO, "\nThe hash table '%s' is of size %d. It has:\n\t%d keys that are never used;\n\ta total of %d elements;\n\tan average linked-list length of %f;\n\tand a maximum linked-list length of %d. \n\n", hash_table_name, HASHSIZE, num_NULL, total_elements, avg_num, max_num); 
+	vpr_printf(TIO_MESSAGE_INFO, "\nThe hash table '%s' is of size %d. It has:\n\t%d keys that are never used;\n\ta total of %d elements;\n\tan average linked-list length of %.1f;\n\tand a maximum linked-list length of %d. \n\n", hash_table_name, HASHSIZE, num_NULL, total_elements, avg_num, max_num); 
 }
